organizes content guidelines utils content into a folder
