+++ next/pages/guidelines/color/usage/icons.tsx
@@ -5,9 +5,9 @@ import { ContentPage } from '../../../../components/mdx/mdx';
 import { H2, H3, LI, P, UL } from '../../../../components/mdx/components';
 import getLayoutProps from '../../../../utils/get-layout-props';
 import ExampleBox from '../../../../components/example-box';
-import UsageCategory from './usage-categories';
-import ColorUsageNav from './color-usage-nav';
-import { Usage, Image } from '../../../../utils/color-usage-types';
+import UsageCategory from '../../../../utils/guidelines/color/color-usage-categories';
+import ColorUsageNav from '../../../../utils/guidelines/color/color-usage-nav';
+import { Usage, Image } from '../../../../utils/guidelines/color/color-usage-types';
 
 import intro from '../../../../images/pages/guide/product/color/usage/icons/intro.png';
 import neutral from '../../../../images/pages/guide/product/color/usage/icons/neutral.png';
